refactor: break down TrackerComponents into modular files
- Split TrackerComponents.tsx into focused component files:
  - components/RoundHeaderComponents.tsx
  - components/DurationComponents.tsx
  - components/EffectsComponents.tsx
  - components/TriggersComponents.tsx
  - components/SessionComponents.tsx
- Reduced complexity and file size issues identified by Codacy
- Maintained all existing functionality and test coverage
- Improved maintainability with smaller, focused components
